==> notebooks/1_Data Preprocessing.ipynb <==
#import
import fastf1
import os
import pandas as pd

#store data
fastf1.Cache.enable_cache(r"\cache")
#pick 2021-2024 data
years = [2021, 2022, 2023, 2024]
laps_data = {}
cleaned_laps_data = {}
preprocessed_laps = []

#detch all the race data from 2021-2024
for year in years:
    print(f"Loading Abu Dhabi {year} GP session...")
    session = fastf1.get_session(year, 'Abu Dhabi', 'R')
    session.load()
    laps_data[year] = session.laps


print("\nAll sessions loaded.")

## Data Cleaning

#go through all fetched data for cleaning
for year in years:
    print(f"Cleaning laps for {year}")
    df = laps_data[year].copy()

    #remove no lap time laps
    df = df[df['LapTime'].notna()]

    #keep accurate laps according to dataset
    df = df[df['IsAccurate'] == True]

    #if tires compound missing j use label as unkown
    df.loc[:, 'Compound'] = df['Compound'].fillna('UNKNOWN')

    cleaned_laps_data[year] = df
    print(f"   -> Cleaned laps: {df.shape[0]} rows\n")


print(f"\nAll laps cleaned.\n") 

## Data Preprocessing

for year in years:
    print(f"Preprocessing cleaned laps for {year}\n")
    df = cleaned_laps_data[year].copy()

    #new column for laptime in seconds
    df['LapTimeSeconds'] = df['LapTime'].dt.total_seconds()

    #append year column
    df['Year'] = year

    preprocessed_laps.append(df)

print("Preprocessing complete.\n")

#create directory for data storage
output_dir = r"C:\data"

#make a dataframe with all the preprocessed laps
combined_df = pd.concat(preprocessed_laps, ignore_index=True)

#keeping only the needed columns
combined_df = combined_df[[
    'Year', 'Driver', 'Stint', 'LapNumber', 'Compound',
    'LapTimeSeconds', 'TrackStatus', 'PitInTime', 'PitOutTime'
]]

#save the data to a csv
output_path = os.path.join(output_dir, "lap_time.csv")
combined_df.to_csv(output_path, index=False)

print(f"Preprocessed data saved to: {output_path}")
