==> notebooks/3_Feature Engineering.ipynb <==
import pandas as pd

#load data
laps_df = pd.read_csv("C:\data/lap_time.csv")

print("Loaded data")

laps_df.head()

#split training and testing roughly 75/25
train_laps = laps_df[laps_df['Year'].isin([2021, 2022, 2023])].copy()

test_laps = laps_df[laps_df['Year'] == 2024].copy()


#extract features and working on feature selection
def extract_clean_features(laps, total_laps=58):
    #list to add all the features needed to train the model
    features = []
    #each driver on a certain tire 
    for (driver, stint), group in laps.groupby(['Driver', 'Stint']):
        group = group.sort_values('LapNumber')
        #best lap time on that stint
        baseline = group.iloc[0]['LapTimeSeconds']
        #where the current stint begans
        start_lap = group['LapNumber'].min()

        for _, row in group.iterrows():
            lap_number = row['LapNumber']
            #how long has the driver been on the said tyre
            laps_in_stint = lap_number - start_lap + 1
            #tyre health
            degradation = row['LapTimeSeconds'] - baseline
            #pitted or not
            pit_event = int(pd.notna(row['PitInTime']) or pd.notna(row['PitOutTime']))
            #safety cars and yellow flags
            flagged = 0 if row['TrackStatus'] == 1 else 1
            #add all the features to the list
            features.append({
                'Driver': row['Driver'],
                'LapNumber': lap_number,
                'Compound': row['Compound'],
                'LapsInStint': laps_in_stint,
                'Degradation': degradation,
                'PitEvent': pit_event,
                'Flagged': flagged,
                'LapPercent': lap_number / total_laps,
                'LapTimeSeconds': row['LapTimeSeconds']
            })

    return pd.DataFrame(features)


#extract only the needed features from our df
train_features = extract_clean_features(train_laps)
test_features = extract_clean_features(test_laps)

extract_clean_features(train_laps)

extract_clean_features(test_laps)

#store it to a pickle file so its easier to access
train_features.to_pickle("C:\data\train_features.pkl")

test_features.to_pickle("C:\data\test_features.pkl")
